requirements.txt
asgiref==3.8.1
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
colorama==0.4.6
Django==5.1
django-filter==24.3
djangochannelsrestframework==1.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gitdb==4.0.11
gitflow==0.5.1
GitPython==3.1.43
gunicorn==23.0.0
h11==0.14.0
materialdash==0.0.21
msgpack==1.1.0
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
redis==5.2.0
smmap==5.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.32.0
websockets==13.1
